Feat: 마이프로필 디자인 수정
