+++ package.json
@@ -11,6 +11,7 @@
         "@types/jquery": "^3.5.16",
         "@types/node": "^18.13.0",
         "@types/react": "^18.0.27",
+        "@types/react-beautiful-dnd": "^13.1.3",
         "@types/react-cookies": "^0.1.0",
         "@types/react-dom": "^18.0.10",
         "@types/react-redux": "^7.1.25",
@@ -36,15 +37,18 @@
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^13.5.0",
         "@types/react-slick": "^0.23.10",
+        "@types/swiper": "^6.0.0",
         "antd": "^5.1.7",
         "axios": "^1.3.2",
         "bootstrap": "^5.2.3",
         "face-api.js": "^0.22.2",
         "jwt-decode": "^3.1.2",
         "react": "^18.2.0",
+        "react-beautiful-dnd": "^13.1.1",
         "react-bootstrap": "^2.7.0",
         "react-cookies": "^0.1.1",
         "react-dom": "^18.2.0",
+        "react-draggable": "^4.4.5",
         "react-icons": "^4.7.1",
         "react-minimal-side-navigation": "^1.9.2",
         "react-modal": "^3.16.1",
@@ -66,6 +70,7 @@
         "slick-carousel": "^1.8.1",
         "socket.io-client": "^4.6.0",
         "styled-components": "^5.3.6",
+        "swiper": "^9.0.5",
         "web-vitals": "^2.1.4",
         "webm-to-mp4": "^1.0.0"
     },
