+++ src/App.tsx
@@ -2,15 +2,14 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes, Link } from 'react-router-dom';
 import 'bootstrap/dist/css/bootstrap.css';
 import Auth from './views/Auth';
-import LibraryContent from './views/Contents/LibraryContent/LibraryContent';
+import LibraryContent from './views/Contents/AlbumListContent/AlbumListContent';
 import VideoContent from './views/Contents/VideoContent/VideoContent';
 import CamContent from './views/Contents/CamContent/CamContent';
 import RemoteCamera from './views/Camera/RemoteCamera';
 import FriendContent from './views/Contents/MyContent/FriendContent';
 import './styles/App.css';
-import { createGlobalStyle } from 'styled-components';
-import AlbumSinglePage from './views/Album/AlbumSinglePage';
-import AlbumMultiContent from './views/Contents/AlbumContent/AlbumMutiContent';
+import SingleVideo from './views/Album/SingleVideo';
+import SingleAlbumContent from './views/Contents/AlbumContent/SingleAlbumContent';
 import Login from './views/Login/Login';
 import Parallax from './views/Main/Parallax';
 
@@ -26,9 +25,9 @@ function App() {
                     <Route path="/auth" element={<Auth />} />
                     <Route path="/remote" element={<RemoteCamera />} />
                     <Route path="/video" element={<VideoContent />} />
-                    <Route path="/video/:cardId" element={<AlbumSinglePage />} />
+                    <Route path="/video/:cardId" element={<SingleVideo />} />
                     <Route path="/album" element={<LibraryContent />} />
-                    <Route path="/album/:albumId" element={<AlbumMultiContent />} />
+                    <Route path="/album/:albumId" element={<SingleAlbumContent />} />
                 </Route>
             </Routes>
         </Router>
