+++ src/components/Header/Headers.css
@@ -1,0 +1,95 @@
+.navigation__wrapper {
+    width: 100%;
+    height: 100px;
+    z-index: 9999;
+    background-color: #fff;
+    font-family: IBMPlexSansKR-Regular;
+}
+s .menu__list {
+    display: flex;
+    float: right;
+    /* justify-content: center; */
+}
+
+@media (max-width: 1099px) {
+    .burger__menu,
+    .x__menu {
+        display: block;
+        float: right;
+        cursor: pointer;
+        top: 50px;
+        z-index: 9998;
+    }
+
+    .burger__menu > div,
+    .x__menu > div {
+        width: 25px;
+        height: 3px;
+        background-color: black;
+        margin: 5px;
+        transition: all 0.3s ease;
+        justify-content: bottom;
+        z-index: 9998;
+    }
+
+    .x__menu > .burger_line1 {
+        transform: rotate(-45deg) translate(-5px, 6px);
+    }
+
+    .x__menu > .burger_line2 {
+        opacity: 0;
+    }
+
+    .x__menu > .burger_line3 {
+        transform: rotate(45deg) translate(-5px, -6px);
+    }
+
+    .menu__box__visible {
+        width: 130px;
+        height: 100%;
+        position: fixed;
+        text-align: center;
+        right: 0;
+        top: 100px;
+        background-color: white;
+        /* box-shadow: 2px 0px 1px rgba(44, 44, 44, 0.137); */
+    }
+
+    .menu__box__hidden {
+        display: none;
+    }
+
+    .menu__list {
+        position: relative;
+        top: 50px;
+        left: 0px;
+        flex-direction: column;
+    }
+
+    .menu__item {
+        margin: 15px 0;
+    }
+}
+
+@media (min-width: 1100px) {
+    .navigation__wrapper {
+        flex-direction: row;
+        align-items: center;
+        justify-content: center;
+    }
+
+    .burger__menu,
+    .x__menu {
+        display: none;
+    }
+
+    .menu__box {
+        width: 50%;
+        height: fit-content;
+    }
+
+    .menu__list {
+        flex-direction: row;
+        align-items: center;
+    }
+}
