--- src/components/Sidebar/Profile/Profile.tsx
@@ -1,20 +1,0 @@
-// import { Avatar } from 'antd'
-// import React, { useState, useRef } from 'react'
-
-
-export default function Profile() {};
-//     const [image, setImage] = useState('https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png')
-//     const fileInput = useRef(null)
-
-//     const handleClick = () => {
-//         fileInput.current.click()
-//     }
-//     <input type='file' style={ }
-//         return (
-//             <>
-//                 <Avatar src={image} style={{ margin: '20px', size: 200 }} onClick={handleClick} />
-
-
-//             </>
-//         )
-// }
