+++ src/views/Contents/AlbumContent/SingleAlbumContent.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SingleAlbum from '../../Album/SingleAlbum';
+
+export default function AlbumContent() {
+    return (
+        <div>
+            <SingleAlbum />
+        </div>
+    );
+}
