+++ src/views/Contents/AlbumContent/SingleVideoContent.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import SingleVideo from '../../Album/SingleVideo';
+
+export default function AlbumContent() {
+    return (
+        <div>
+            <SingleVideo />
+        </div>
+    );
+}
