+++ src/views/Contents/AlbumListContent/AlbumListContent.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import AlbumList from '../../AlbumList/AlbumList';
+
+export default function AlbumListContent() {
+    return (
+        <div>
+            <AlbumList />
+        </div>
+    );
+}
