+++ src/views/Contents/MyContent/EditContent.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import Edit from '../../MyPage/Edit';
-import Sidebar from '../../../components/Sidebar/Sidebar';
+
 import styled from 'styled-components';
 import Header from '../../../components/Header/Header';
-import Footer from '../../../components/Footer/Footer';
 
 const Center = styled.div`
     height: 100vh;
@@ -19,10 +18,8 @@ export default function EditContent() {
                 <hr />
             </span>
             <Center>
-                <Sidebar />
                 <Edit />
             </Center>
-            <Footer />
         </div>
     );
 }
