+++ src/views/Contents/MyContent/FriendContent.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Friend from '../../MyPage/Friend/Friend';
-import Sidebar from '../../../components/Sidebar/Sidebar';
+
 import styled from 'styled-components';
 import Header from '../../../components/Header/Header';
-import Footer from '../../../components/Footer/Footer';
+
 import Edit from '../../MyPage/Edit';
 
 const Center = styled.div`
@@ -20,11 +20,9 @@ export default function FriendContent() {
                 <hr />
             </span>
             <Center>
-                <Sidebar />
                 <Edit />
                 <Friend />
             </Center>
-            <Footer />
         </div>
     );
 }
