+++ src/views/Contents/VideoContent/VideoContent.tsx
@@ -1,24 +1,22 @@
 import React from 'react';
 import styled from 'styled-components';
-import Sidebar from '../../../components/Sidebar/Sidebar';
 import Video from '../../Video/Video';
 import Header from '../../../components/Header/Header';
-import Footer from '../../../components/Footer/Footer';
 
 const StyleContent = styled.div`
-    background-color: #f1f5f9;
+    /* background-color: #f1f5f9;
     font-family: IBMPlexSansKR-Regular;
     box-sizing: border-box;
     display: flex;
     height: auto;
     margin-top: -16px;
-    text-align: center;
+    text-align: center; */
 `;
 const VideoContent = styled.section`
-    margin: auto;
+    /* margin: auto;
     /* overflow: hidden; */
-    display: block;
-    height: auto;
+    /* display: block; */
+    /* height: auto; */
 `;
 
 export default function CamContent() {
@@ -29,7 +27,6 @@ export default function CamContent() {
                 <hr />
             </span>
             <StyleContent>
-                <Sidebar />
                 <VideoContent>
                     <Video />
                 </VideoContent>
@@ -37,7 +34,6 @@ export default function CamContent() {
             <span>
                 <hr style={{ marginTop: '-1px' }} />
             </span>
-            <Footer />
         </div>
     );
 }
