+++ src/views/MyPage/Friend/index.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Document</title>
+        <link rel="stylesheet" type="text/css" href="./friend.css" />
+    </head>
+    <body>
+        <div class="card">
+            <div class="imgBx">
+                <img src="" />
+            </div>
+            <div class="content">
+                <div class="details">
+                    <h2>kimkimkim<br /><span>nick name is?</span></h2>
+                    <div class="data">
+                        <h3>342<br /><span>Posts</span></h3>
+                        <h3>342<br /><span>Followers</span></h3>
+                        <h3>342<br /><span>Following</span></h3>
+                    </div>
+                    <div class="actionBtn">
+                        <button>Follow</button>
+                        <button>Message</button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </body>
+</html>
